docs(Button): move to components folder
