+++ src/components/Button/docs/ButtonExtendedExample.st.css
@@ -1,5 +1,5 @@
 :import {
-  -st-from: "../../src/components/Button/Button.st.css";
+  -st-from: "../Button.st.css";
   -st-default: TPAButton;
 }
