+++ src/components/Button/docs/ButtonExtendedExample.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {ButtonProps, Button} from '../../src/components/Button';
+import {ButtonProps, Button} from '../Button';
 import extendedStyles from './ButtonExtendedExample.st.css';
 
 export const ButtonExtendedExample: React.FunctionComponent<ButtonProps> = (props) =>
