+++ src/components/Button/docs/examples.tsx
@@ -8,36 +8,38 @@ import * as ExtendedCSSRawSource from '!raw-loader!./ButtonExtendedExample.st.cs
 import * as ExtendedCSSRawSourceWithStyleParams from '!raw-loader!./ButtonExtendedWithStyleParamsExample.st.css';
 import { ButtonExtendedExample } from './ButtonExtendedExample';
 import { ButtonExtendedWithStyleParamsExample } from './ButtonExtendedWithStyleParamsExample';
-import { MockSettings } from '../helperComponents/MockSettings';
+import { MockSettings } from '../../../../stories/helperComponents/MockSettings';
 
 export const Examples = () => (
   <div>
-      <div className="tpa-container">
-          <CodeExample
-              title="Button Extended"
-              code={[
-                  '//.st.css',
-                  ExtendedCSSRawSource,
-                  '',
-                  '//.tsx',
-                  ExtendedRawSource,
-              ].join('\n')}
-          >
-              <ButtonExtendedExample>some text</ButtonExtendedExample>
-          </CodeExample>
-      </div>
+    <div className="tpa-container">
+      <CodeExample
+        title="Button Extended"
+        code={[
+          '//.st.css',
+          ExtendedCSSRawSource,
+          '',
+          '//.tsx',
+          ExtendedRawSource,
+        ].join('\n')}
+      >
+        <ButtonExtendedExample>some text</ButtonExtendedExample>
+      </CodeExample>
+    </div>
     <div className="tpa-container">
       <CodeExample
         title="Button with style params"
         code={[
           '//.st.css',
-            ExtendedCSSRawSourceWithStyleParams,
+          ExtendedCSSRawSourceWithStyleParams,
           '',
           '//.tsx',
-            ExtendedRawSourceWithStyleParams,
+          ExtendedRawSourceWithStyleParams,
         ].join('\n')}
       >
-        <ButtonExtendedWithStyleParamsExample>some text</ButtonExtendedWithStyleParamsExample>
+        <ButtonExtendedWithStyleParamsExample>
+          some text
+        </ButtonExtendedWithStyleParamsExample>
       </CodeExample>
     </div>
     <MockSettings
