+++ src/components/Button/docs/index.story.tsx
@@ -1,21 +1,20 @@
 import * as React from 'react';
-import {Button, PRIORITY, SIZE} from '../src/components/Button';
-import * as ButtonSource from '!raw-loader!../src/components/Button/Button.tsx';
-import {Examples} from './Button';
+import {Button, PRIORITY, SIZE} from '..';
+import {Examples} from './examples';
 
 export default {
   category: 'Components',
   storyName: 'Button',
   component: Button,
-  source: ButtonSource,
-  componentPath: '../src/components/Button/Button.tsx',
-  componentProps: () => ({
+  componentPath: '../Button.tsx',
+
+  componentProps: {
     'data-hook': 'storybook-Button',
     children: 'Book Now',
     priority: PRIORITY.basic,
     size: SIZE.medium,
     fullWidth: false,
-  }),
+  },
   exampleProps: {
     priority: Object.keys(PRIORITY).map(key => PRIORITY[key]),
     size: Object.keys(SIZE).map(key => SIZE[key]),
